[Task-11] Create survey adapter and lazy load when fetch survey list
